+++ src/components/common/index.ts
@@ -1,4 +1,5 @@
 export { default as Button1 } from './Button1'
+export { default as CountdownCard } from './CountdownCard'
 export { default as Button2 } from './Button2'
 export { default as Button3 } from './Button3'
 export { default as ArrowIcon } from './icons/ArrowIcon'
