+++ src/components/home/HomeRecruitSection.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from 'react'
 import { motion, AnimatePresence } from 'motion/react'
-import { Button1 } from '@/components/common'
+import { Button1, CountdownCard } from '@/components/common'
+
 import { RECRUIT_SECTION_DATA, RECRUIT_TARGET_DATE } from '@/constants/home/recruitSection'
 import { useHomeRecruitAnimation } from '@/hooks/home/useHomeRecruitAnimation'
 
@@ -17,34 +18,6 @@ function calcTimeLeft(): TimeLeft {
   }
 }
 
-function pad(n: number) {
-  return String(n).padStart(2, '0')
-}
-
-function CountdownCard({ label, value }: { label: string; value: number }) {
-  const display = pad(value)
-  return (
-    <div className="relative flex flex-1 flex-col items-center gap-3 overflow-hidden rounded-[32px] px-8 pb-7 pt-9 text-center">
-      <div aria-hidden="true" className="absolute inset-0 rounded-[32px] bg-black" />
-      <div className="absolute inset-0 rounded-[32px] shadow-[inset_4px_8px_20px_0px_rgba(255,255,255,0.3)]" />
-      <p className="relative text-[20px] font-medium leading-normal tracking-[-0.4px] text-white/60">{label}</p>
-      <div className="relative h-[90px] w-full overflow-hidden">
-        <AnimatePresence mode="popLayout">
-          <motion.p
-            key={display}
-            className="absolute inset-x-0 text-center text-[60px] font-bold leading-normal tracking-[-1.2px] text-primary-light-02"
-            initial={{ y: -32, opacity: 0 }}
-            animate={{ y: 0, opacity: 1 }}
-            exit={{ y: 32, opacity: 0 }}
-            transition={{ duration: 0.25, ease: 'easeOut' }}
-          >
-            {display}
-          </motion.p>
-        </AnimatePresence>
-      </div>
-    </div>
-  )
-}
 
 function SheenButton({ href }: { href: string }) {
   const [hovered, setHovered] = useState(false)
